feat: Devide characterServices
