+++ GhostInTheShell.Modules.InfraStructure/WellknownConfigurationSecsions.cs
@@ -1,0 +1,14 @@
+﻿namespace GhostInTheShell.Modules.InfraStructure
+{
+    public class WellknownConfigurationSecsions
+    {
+        public const string FactoryType = "Shell:FactoryType";
+        public const string ShellName = "Shell:Name";
+
+        public const string RemoteTableRoot = "Shell:Remote:TableRoot";
+        public const string LocalTableRoot = "Shell:Local:TableRoot";
+
+        public const string RemoteMaterialRoot = "Shell:Remote:MaterialRoot";
+        public const string LocalMaterialRoot = "Shell:Local:MaterialRoot";
+    }
+}
