--- GhostInTheShell.Modules.Shell/IColorable.cs
@@ -1,10 +1,0 @@
-﻿using GhostInTheShell.Modules.InfraStructure;
-
-namespace GhostInTheShell.Modules.Shell
-{
-    public interface IColorable
-    {
-        bool ChangeColor(Hsl hslColor);
-        void ChangeDefaultColor();
-    }
-}
