+++ GhostInTheShell.Modules.Shell/ShellWindow.xaml
@@ -13,7 +13,7 @@
         AllowsTransparency="True"
         Width="{Binding ShellSize.Width}" Height="{Binding ShellSize.Height}">
 
-    <!--<ContentControl prism:RegionManager.RegionName="{x:Static infra:WellknownRegionNames.ShellViewRegion}"
+    <ContentControl prism:RegionManager.RegionName="{x:Static infra:WellknownRegionNames.ShellViewRegion}"
                     Width="{Binding ShellSize.Width}"
-                    Height="{Binding ShellSize.Height}"/>-->
+                    Height="{Binding ShellSize.Height}"/>
 </Window>
