+++ GhostInTheShell.Modules.Shell/Views/ShellView.xaml
@@ -9,7 +9,6 @@
              d:DesignHeight="450" d:DesignWidth="800"
              d:DataContext="{d:DesignInstance Type={x:Type localViewModel:ShellViewModel}}"
              xmlns:prism="http://prismlibrary.com/"
-             
              Width="{Binding ShellSize.Width}" Height="{Binding ShellSize.Height}">
 
     <Image Width="{Binding ShellSize.Width}" Height="{Binding ShellSize.Height}" Source="{Binding ShellSource}"/>
